fix: cambiar tipos de variable a Interger para solucionar problemas de mapeo de la BD

+++ src/main/java/com/example/nutriplan_backend/controller/ActividadFisicaController.java
@@ -23,7 +23,7 @@ public class ActividadFisicaController {
 
     // Obtener por ID
     @GetMapping("/{id}")
-    public ResponseEntity<ActividadFisica> obtenerPorId(@PathVariable Long id){
+    public ResponseEntity<ActividadFisica> obtenerPorId(@PathVariable Integer id){
         ActividadFisica consulta = actividadFisicaService.obtenerPorId(id);
         return ResponseEntity.ok(consulta);
     }
